Add various ways to visualize corner geometry
bezier-intersection.html is a proving ground for comparing the bezier-based
superellipse with the iterative version, and for validating the bezier splitting
logic required for inner edges.

corner-geometry.html presents the math required to compute inner superellipse
curves, showing the offset corner points and control points.

shape-renderer.html is a playground for the various CSS corner treatments,
and computes the full inner and outer bezier curves.